Updated the list of possible interactive interfaces in the BlueSkyBrainstorming file.
I had a conversation with my mother about what she might be interested to see in a theme park, and although she's heard from me as I've worked on the theoretical side of the project, she doesn't seem to have really internalized what it was about, so I got an array of answers, some of which were wholly unhelpful, but also many that were intriguing in the ways they fell slightly outside the sphere of what I've been considering. In particular, among the ideas my mother suggested, she mentioned consuming various IP specific foods (such as from Harry Potter, Lord of the Rings, or Alice in Wonderland), wearing specifically iconic clothing (princess dresses, cultural fashions, etc), and some specific embodied experiences, such as falling down the rabbit hole in Alice in Wonderland. Although these might not be technologically interactive interfaces at first blush, they do all present some sense of acts (eating/drinking/dressing/falling) with sensory feedback that further reinforce immersion in the experience. I'm including them in my list because I think it provokes some interesting lines of thought for where one draws the line on what is "interactive".

For instance, we might not think putting on a shirt is interactive, but is getting kitted out in armour by a squire at a renaissance fair interactive? Is an Iron-Man helmet with automated moveable visor interactive? Does piloting a mech count as wearing it? Similarly, eating and drinking might not be "interactive" in an of themselves, per se, but bars such as Disney's Trader Sam's Enchanted Tiki Bar have special effects that trigger in the bar when you order specific drinks, which certainly feels interactive, and 4e Mur here in Montreal is a bar where you can solve a mystery based on clues that come with ordering specific drinks. In this fashion, it becomes clear that "participatory" activities do have robust design opportunities to be made fully interactive, and so I want to put some more thought into this at some point.