Created and started drafting a document called "WhatIsShooting".
This document is to hopefully explore some ideas around shooting as a concept, both aesthetically and mechanically. Though it is easy to criticize it, morally as well as creatively, I do think there is perhaps some greater diversity to it than I've given it credit for. I want to use this document as a way to explore and elaborate the idea, not neccesarily to use it, but to at least leave the possibility open should I come across something really novel or intriguing. Even considering the number of interactions that map onto "hitting a target at a distance" by some vague definition, i think merits closer examination.