Added some notes on the specific video games listed in the BlueSkyBrainstorming document
Blocked out some of the themes and interactions I'm seeing in top watched video games.

Unsurprisingly, shooting continues to be a recurring theme, though once again I'm struck by the relative diversity of shooting. In particular, my eye was caught by some of the variety in gameplay modes, class systems that feature prominently in many games, and the variety of additional common features such as ults. I'm also starting to think about how, for many of these games, even healing is sometimes executed out as a "gun", such as Overwatch's Mercy, or, to pull an old reference, the Medic from TF2. I also fell down a rabbit hole examining Fortnite and it's unusual features of building, often leveraged competitively, and dancing, both as a mode of expression, and game mechanic in taco time and boogie bombs.

As much as I was initially skeptical of shooting as a mechanic, I'm increasingly wondering if part of the issue isn't so much the prevalence of shooting, so much as it's convergeance on a specific execution of shooting. I find myself wondering if it might be possible to design, say, a laser tag system that achieves some of the fantasy to which games like Fortnite, Overwatch, Valorant, or Apex Legends cater to– with team strategy, and unique abilities. I think this might be worth exploring further, in part because it's interesting, but also in the interest of just starting to actually make something.

Which brings me to my other realization of just how transformative this journaling method is for how I work. I'm so used to a slow percolation and intermittent tinkering with both materials and ideas that if left to my own devices, I'm not usually end on, or even with, a self-contained chunk of work. This has led, in a few cases, to me starting something, and then not actually journalling about it until I finished it a week or two later, or, for fear that a do the above, entirely avoiding working at all, lest I lose the freshness of the ideas when journalling. I'd initially thought I was just purely too busy and got distracted from working on this project, but I'm starting to realize that if I was not using this method of tracking progress, I probably would have been making small, incremental progress in between other work. Instead, I simply didn't do anything for three weeks, then didn't do anything for a week and a half again. I'm going to try pushing through this going forward but it's both concerning for productivity, and for the validity of how I'm using this journalling method, because it changes how I design in such a radical fashion.