I put together a preliminary list of affordances in the Splatoon franchise.
I spent some time watching tutorials, tip videos, and reading wikis to get a sense of the basic mechanics of Splatoon and the general ways those mechanics tend to be used. What immediately strikes me is that a core affordance of the game, swimming, is implicated in ammo, life, visibility, and mobility in ways that are not immediately obvious in how they might be rendered tangible. I can sorta imagine a system of swapping between arbitrary states, but the fast-paced mobility options seem to be a pretty integral part of play. Maybe this could be incorporated in some way into like, variable-speed rollerskates, though that still does not address the question of visibility.

I'm also thinking about how the inking of territory might be done. Obviously this would require either projectors or screens. Projectors might be the easier option for my purposes, both for cost and safety of the tech. Targeting might be trickier, but can probably be done with some kind of machine vision and laser pattern? I'll have to look into possible tech solutions for this.

Consequently adding two items to the to-do list:
- Look into mobility tech
- Look into targeting tech