Created a document to start brainstorming ways to execute the affordances of Splatoon
Currently focusing on the gameplay elements of painting the environment, eliminating opponents, and increasing mobility. I think the technical way to achieve the first two might actually be pretty straightforward with a modified conventional laser tag arrangement, albeit with considerable tech requirements, so that does preoccupy my mind. On some level, I'm not super concerned with the economic viability of the experience, but I do think I have to be mindful to the potential concerns of cost and maintenance if the system's demands are far too heavy, simply because a plan without practicality doesn't actually go very far.

Furthermore, I also find myself thinking about the pace of Splatoon. It seems both fast-paced and fluid. Various mechanics allowed for a game that's as much about mobility as it is about shooting, and so I feel like that pace adds to the appeal of the game. Achieving that mobility is an open question though, as conventional laser tag already typically asks its patrons to not even run to avoid injury. This is in large part because conventional laser tag uses darkened mazes, creating both low visibility and heightened obstacles, but I wonder if this potentially pushes me towards a more open-plan play field.

A few factors push me to consider an open plan:
- The aforementioned reduction of dangerous obstructions, so as to facilitate greater mobility.
- Reducing the number of projectors and cameras required, as they could number fewer and be positioned in a top-down fashion.
- Splatoon doesn't count vertical surfaces as painted territory for the purpose of the game objective, so apart from providing cover, vertical obstacles may actually be incidental to the gameplay anyways.
These three factors taken together seem to provide pretty compelling and comprehensive reasoning to steer towards a more open plan, so I'll start thinking in that direction for now.