Initiated the prototyping stage of the project
This update involved a few separate but related moving parts, all pertaining to moving the project from the bluesky brainstorming stage to the prototyping of a concrete experience.

First, previous documents were moved into a Bluesky folder and a Prototyping folder was created to hold iterations of the prototype, and the to-do list was updated to reflect the shift in project tasks. This is all pretty boilerplate.

Second, I've started working on the first prototype, leveraging the conceptual tech from The Graffiti Research Lab's blitzTag platform, which basically uses a camera, projector, and laser pointer to measure relative luminance as the input for output graphics. I was able to get my hands on a conventional laptop presentation laser point, much lower in brightness than the high-power green pointers used by the Graffiti Research Lab, but even so, I'm having some very preliminary success at tracking the dot to the exclusion of other objects in the shot. I have not yet been able to test it with a projector running, as I'm currently travelling, but I do think this is a promising tentative first step towards a playable prototype and should certainly be testable mid-next-week once I'm home. I do foresee there being issues with dialling in the exact network for extracting the laser dot, as the projector itself will produce light, of course, but so far I've been testing in a lit room and as long as the camera isn't pointed directly at an ambient light source, it does seem possible to distinguish between laser and ambient light. If this does prove to be a problem, I can always see about obtaining a more powerful laser, but that will have to be a question of safety against laser power. The other distinct question of scalability will also come into play if I can get a desktop prototype to work, but one that doesn't scale properly to the size of a play arena. In Graffiti Research Lab's blitzTag, they use a reasonably high-power laser because they're "painting" at a distance on the side of a building, so that maybe gives me a sense of the benchmark I'm designing against, but then again, their project seems to be a little bit older (references PS3 camera, page comments date from early 2010s) so I may also benefit from simply better hardware.

The final point I want to touch on with this update is returning to my proposed design principles, as the whole point of this project is to be enacting them and I feel like I've already strayed from them once previously. As I move towards a prototype, I think I'm keeping at least partially in-line with principles I and II regarding bottom-up design and iconic affordances. Principles III, IV, and V are still open questions I will have to address as I move forward with higher fidelity prototypes, but I am noticing that my previous uncertainty regarding the simulation of mobility mechanics from Splatoon will likely run directly into at least principle IV regarding feedback, and possibly principle V on dissimulation. Now that I'm actually making a thing, I'll likely have to circle back to these principles much more often to ensure I'm balancing practical considerations against the theory.