Added some thoughts on mobility mechanics to the v001 prototype description.
I'm probably not building out mobility for the first prototype, but I'm already bouncing ideas around. So far I've really been fixated on if there's a way to get skates of some kind (roller or ice) to work, but it's still very unclear to me these would be safe, or even effective at replicating the specific affordances of Splatoon, thereby running afoul of design principle II. "Sliding" gear borrowed from haunted attractions, whereby participants can slide on their hands and knees might prove a useful direction, and this again has some safety concerns, but perhaps comes closer to replicating the feeling of a "squid submersed in ink" that happens in an actual Splatoon game. However, sliding is kinda a specific skill, and while not entirely impossible to pick up, the learning curve might pose its own challenges against design principle IV regarding competency. Obviously I'm not going to be solving this problem tonight, and I've got plenty of time to do it, but this does feel like it's going to be a much tougher nut to crack than the shooting aspect of the experience.