Got a very quick basic functionality for shooting up and running.
Despite travelling, was able to use a basic laser pointer to get the basic functionality up and working. Should be easier to test once I'm in an environment where I can control the setup more precisely, but next step will be to test it with a projector, to see if the luminance of the projection can be balanced against the laser.

Will also need to think about how to discern between the lasers of different weapons (either for functionality or to discern team colour). Trying to identify a laser symbol might be a bit tricky, but I wonder if I could instead use frequency of flashing to do it. This would obviously be a bit more complicated and would require creating a modified programmable laser, but the hardware side shouldn't be too difficult with the help of an arduino board. Detecting frequency of flash on the laser is theoretically easy (it's how remote controls worth with IR light) but I'll have to do some research into precisely how to enact that in TouchDesigner.