No file updates, but conducted a very quick and dirty test of the setup depicted in Diagram.jpg.
I was pleasantly surprised at how well it work. I will need to test again in a more controlled environment (large flat surface, adjustable lighting) but even in a noisy environment (just a wall of my bedroom) I was getting pretty good discrimination on the laser pointer's position to the exclusion of anything else. It's definitely still a low-light environment, but I think I might be able to get away with low-but-still decent lighting which improves concerns of safety by some imprecise margin.