Added a "splatter" shape to where the laser shoots.
This is an aesthetic question issue, but I think it's part of what creates the satisfying "splat" of splatoon, as well as games with similar dynamics such as Super Mario Sunshine. Right now the splat is a pretty uniform shape so I still want to experiment with alternative ways to make the splat a bit more random. Also sound effects and some real texturing for the splat would probably go some ways to adding life. Finally I want to figure out a little animation so you see the splat spread as it "hits" the surface. I think apart from the aesthetics, this would also help with making the "shooting" more intuitive since you otherwise can't see your projectile.