Added a noise function that rotates the "splat" shape to give it some more natural variety.
It's basically fully random right now so to tends to lay down a lot of superimposed splats for any given shot, obscuring the actual shape, which has somewhat more natural contours in that they're rough and messy, but is also making me realize I'm not super happy with the splat itself and I don't think this random pinwheeling will work if I want to animate the splat.

At this stage I'm actually wondering if it's entirely a mistake to be putting this time into creating the splat as a 2D graphic, rather than just exploring options for 3D fluid simulation in TouchDesigner. I'm guessing that working with a 3D simulation that I just render onto a 2D surface will probably feel more natural in all regards, rather than trying to replicate the look, feel, and dynamics by hand and math.