Implemented basic framework for scoring.
I know I previously said I might put this off until later, but I realized I could use an analyze node to do the pixel counting I wanted, which would make implementation a snap. Now, unfortunately due to something goofy in what analyze is doing is yielding slight greater than 1 values for aggregate coverage, which of course makes no sense but is probably an artifact of aliased edges on the silhouette. Frankly it's close enough for my purposes (we're talking like 1.011 coverage instead of 1) so I went ahead and implemented scoring, albeit with a bit of extra normalization.

Since the scoring tracks proportion of area covered by each team, as well as uncovered area, the unclaimed ground is what gets cut if the total coverage is reading higher than one. That is to say, I sum both teams coverage and divide by total area. If that number is >1, then I set unclaimed area to 0 and divide the 1 proportionally to the ratio of the two teams coverage relative to each other. If the total <1, then I presume it's accurate, set (1-number) as the unclaimed territory and then divide the number proportionally to the two teams coverage. It's a bit inelegant as implemented but I'm pretty sure it will get the job done in a robust manner. I'll probably want to also implement a GUI to display scoring, as well as a timer for game play, but for now, this was the main hurdle I anticipated with regards to scoring. Everything else is largely cosmetic.