Implemented a score display.
Added a bar along the bottom that shows the proportion of the screen covered by one team or the other. Unfortunately, it seems to be ticking backwards under certain circumstances. Based on how it's coded, I actually think this is not a problem with the display but rather the underlying calculation of area covered. My guess is my convoluted method of normalizing the data is probably throwing a wrench in thing somewhere so I'm going to have to revisit the flow.

I also realized I've polished the visuals a bit but haven't added any sound and I vaguely recall making a note that the sound was important to the experience. There's something about a splat or squelch that I think really makes the inking process more satisfying. In the interest of not overlooking it again, I've added a note about it to Todo.md.

Next steps will probably be to troubleshoot the scoring system, add sound effects, and add numbers to the scoring system so that (when it does work properly) there are also numerical indications of score.