Added numerical display to scoring.
Pretty simple addition, but added a numerical readout to the score display. I think I'll probably want to aesthetically overhaul all of the score display in a later iteration of the prototype, since this current version is so flat and plain, but for now I think this does the job of conveying relevant information on who's in the lead.