Implemented a timer.
Implementing a timer proved easier than I initial anticipated because I remembered TouchDesigner is, after all, designed for time-based work, so I simply set the built-in clock to a time limit of 30 seconds and display the clock time inverted to created a countdown. Again, I'm sure this will change, including the length of the countdown, but the current network should be easy enough to modify once I can start actually testing this prototype as a game.

As a side note for the timer, I debated if I just wanted a raw number of second or a #:## format. Since I'm not using the minutes digit, I'm just going with a raw number countdown, but at a later stage I anticipate wanting to switch over to the minutes + seconds format for longer time limits.