Obtained laser components to build out laser guns.
I purchased some cheap laser components off amazon for the purpose of getting a prototype up off the ground. They nominally operate at 5V, 20mA but on initial test they seemed a bit dull. Since I'll be looking to only flash the dot rather than ever use it continuously, I'll probably experiment with judiciously overdriving the diode to get something a little punchier, without frying it and while still keeping in mind that I don't want to blind any players.

I'll include schematics (and maybe a component list with specs for the sake of capturing non-digital elements of the project?) as I build out the circuit, but I'm thinking an arduino taking a digital input from a "trigger" microswitch, which then simply sends a few quick on-off cycles to the laser+resistor at a prescribed frequency to be identified on the TouchDesigner side. I might actually be able to execute flickering with pulse width modulation but I don't know how convenient that will be to process as a signal on the receiving end. More specifically, I don't know if I'll actually need an identifiable frequency so much as a unique singular interval between two flashes in a sufficiently localized position. Like, I'm pretty sure I'm going to have to spatially isolate laser dots in the video feed anyways, so at that point it might only take the two flashes to establish an identity. I will have to experiment more on the receiving end once I have a functional laser gun prototype.