Built the shooter prototype.
Prototype is in accordance with circuit.png under the v001 folder, where the laser diode is nominally 5v, 20mA, but resistor A in the diagram is initially using a 120 ohm resistor, overdriving the diode during it's brief flashes because it seemed a bit dim otherwise. Shooter is running splatshooter.ino on its arduino uno board and luckily it was simple enough code that i was able to bang it out without any bugs on the first go-through. Currently the shooter flashes once, so there is no team discrimination, so I will have to experiment with adding team discrimination next.

Having tested this initial prototype, the "single shot" effect created by the flash rather than the continuous stream of the previous laser pointer definitely changes the feel of the splat. I know I abandoned animation earlier in the project, but I may want to return to animation give the splat a little more life. Right now it feels a little bit stiff and unresponsive in a way that's hard to describe other than "yes, the laser dot sure makes a static blob shape on the screen." I won't jump right into adding animation first though, since I know I want to add sound anyways. Adding the sound might make the splat feel more "splat-like" so I may not actually need anything further once that's implemented.

Finally, I hadn't initially figured I would prototype a casing for the shooter at this stage, but in it's current state, (see shooter-side and shooter-front) as essentially loose parts held together with tape, I think it's probably neccessary to give some thought to basic ergonomics, even at this stage. Not sure how I'll approach this, either modelling and 3D printing a casing, or just hacking something from recycling bin contents, but it's otherwise such a hassle to have to hold the shooter together while using it, that I think it can't help but influence perception of the affordance.