Voice Note: Added a new folder under my "Process" folder for "Voice Notes" and added my first voice note.
My idea here is I finally gave a voice note a try because I had a thought on the shuttle on my way to campus and couldn't archive it otherwise. I may think about using these more often in future since I have noticed I struggle with design thoughts at inopportune times. For later ease of analysis, here is the transcription of the voice note:

"Okay, trying something a little bit different for this [um] update. It's about 10:50 am on a Friday and I just got of the shuttle to the downtown campus. [um] While on the shuttle my mind just started just sorta wandering thinking about the affordances in Splatoon—[um] and this is probably getting a little bit ahead of my current prototype—but I'm thinking about the different kinds of equipment and weapons that show up in the game. In particular, [uh] I started thinking about their– their giant paint brushes or rollers [um] that are used to paint the environment and I previously had been thinking that these might end up being [uh] sort of laser brooms, if you will, with lasers projected [um] onto the ground just ahead of the broom head for the [um] for the camera to detect, but it's occurring to me that because I am actually just checking for absolute luminosity in the visible frame, I can probably just make it a self-luminous broom head by putting lights inside. [um] As I say, this is probably a little bit ahead of the curve, but I wanted to go ahead and archive this while I had the thought, because I know this is going to be useful at a later stage. "